[AIP-21] Move contrib.operators.spark_sql_operator providers.apache.spark.operators.spark_sql
