+++ docs/operators-and-hooks-ref.rst
@@ -117,7 +117,7 @@ Foundation.
        :mod:`airflow.providers.apache.spark.hooks.spark_sql`,
        :mod:`airflow.providers.apache.spark.hooks.spark_submit`
      - :mod:`airflow.providers.apache.spark.operators.spark_jdbc`,
-       :mod:`airflow.contrib.operators.spark_sql_operator`,
+       :mod:`airflow.providers.apache.spark.operators.spark_sql`,
        :mod:`airflow.contrib.operators.spark_submit_operator`
      -
