+++ scripts/ci/pylint_todo.txt
@@ -67,7 +67,7 @@
 ./airflow/contrib/operators/slack_webhook_operator.py
 ./airflow/contrib/operators/snowflake_operator.py
 ./airflow/providers/apache/spark/operators/spark_jdbc.py
-./airflow/contrib/operators/spark_sql_operator.py
+./airflow/providers/apache/spark/operators/spark_sql.py
 ./airflow/contrib/operators/spark_submit_operator.py
 ./airflow/contrib/operators/sqoop_operator.py
 ./airflow/contrib/operators/ssh_operator.py
